Updated template and converted floats to double in freckles.cpp

+++ C++/problems/freckles.cpp
@@ -11,23 +11,23 @@
 #include <vector>
 #include <cmath>
 
-float round_n(float x, int n = 2) {
+double round_n(double x, int n = 2) {
     return roundf(x * pow(10, n)) / pow(10, n);
 }
 
-float distance(std::pair<float,float> p1, std::pair<float,float> p2){
+double distance(std::pair<double,double> p1, std::pair<double,double> p2){
     return sqrt(pow(p1.first-p2.first,2)+pow(p1.second-p2.second,2));
 }
 
-float shortest_path_alg(std::vector<std::pair<float,float>> points){
+double shortest_path_alg(std::vector<std::pair<double,double>> points){
     // Greedy algorithm to find the shortest path
-    std::vector<std::pair<float,float>> visited;
-    float min_dist = 0;
+    std::vector<std::pair<double,double>> visited;
+    double min_dist = 0;
     int n = points.size()-1;
     visited.push_back(points.back());
     points.pop_back();
     for(int i=0; i<n;i++){
-        std::vector<float> distances(points.size());
+        std::vector<double> distances(points.size());
         for(int j=0; j<points.size();j++){
             distances[j] = distance(visited.back(),points[j]);
         }
@@ -51,13 +51,13 @@ int main() {
     for(int i = 0; i < N; i++){
         int n;
         std::cin >> n;
-        std::vector<std::pair<float, float>> freckles;
+        std::vector<std::pair<double, double>> freckles;
         for(int j = 0; j < n; j++){
-            float x, y;
+            double x, y;
             std::cin >> x >> y;
             freckles.push_back(std::make_pair(x, y));
         }
-        float shrt_path = shortest_path_alg(freckles);
+        double shrt_path = shortest_path_alg(freckles);
         std::cout << round_n(shrt_path) << std::endl;
     }
